vilegal2/requirements.txt
omegaconf==2.0.6
pytorch-lightning>=2.0.0
hydra-core==1.3.2
transformers==4.30.2
datasets==2.12.0
rouge-score==0.1.2
torch>=1.13.0
scikit-learn
pandas
wandb 